Switch to using a DataUpdateCoordinator
This is a big change, without any visible changes to the user.

Internally we're switching from a single entity that polls the API
directly to using a DataUpdateCoordinator to manage the polling.

This will allow us to expose other data from the locks as additional
entities in the future without excessive API use and with consistent
and performant handling of webhook data.